Corrected the issue where only the last register zone had its label changed to white.
o Created 3 separate HWND variables for the the labels for the register zones
o Changed the signature of method PlaceRegisterArea to return an HWND.
o Checking all 3 labels for setting the text color and background color.
o Moved the Black pen creation and clean up to the PlaceRegisterArea method.
o Saving the return value for the PlaceRegisterArea method in their respective variables.